+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+  * `HTTPResponse.ok` is now a property as was planned originally
+
+### Added
+  * `http.HTTPError` exception
+  * `HTTPResponse.raise_for_status` method
+
+### Changed
+  * Split `http.base` module into separate modules like `response.py` and `errors.py`
+  * HTTP exceptions are now importable as `http.errors`
 
 ## [0.1.3] - 2024-10-02
 ### Added